tests: threads: fold customdata tests into main test
Move this small test to the main thread test project.
