--- tests/kernel/threads/custom_data/custom_data_api/CMakeLists.txt
@@ -1,5 +1,0 @@
-include($ENV{ZEPHYR_BASE}/cmake/app/boilerplate.cmake NO_POLICY_SCOPE)
-project(NONE)
-
-FILE(GLOB app_sources src/*.c)
-target_sources(app PRIVATE ${app_sources})
