--- tests/kernel/threads/custom_data/custom_data_api/README
@@ -1,10 +1,0 @@
-$make run
-Running test suite test_customdata_api
-tc_start() - test_customdata_get_set_coop
-===================================================================
-PASS - test_customdata_get_set_coop.
-tc_start() - test_customdata_get_set_preempt
-===================================================================
-PASS - test_customdata_get_set_preempt.
-===================================================================
-PROJECT EXECUTION SUCCESSFUL
