--- tests/kernel/threads/custom_data/custom_data_api/prj.conf
@@ -1,2 +1,0 @@
-CONFIG_ZTEST=y
-CONFIG_THREAD_CUSTOM_DATA=y
