--- tests/kernel/threads/custom_data/custom_data_api/testcase.yaml
@@ -1,3 +1,0 @@
-tests:
-  kernel.threads:
-    tags: kernel threads userspace
