+++ tests/kernel/threads/lifecycle/lifecycle_api/prj.conf
@@ -1,3 +1,4 @@
 CONFIG_ZTEST=y
 CONFIG_THREAD_MONITOR=y
 CONFIG_HEAP_MEM_POOL_SIZE=256
+CONFIG_THREAD_CUSTOM_DATA=y
